+++ expecter/__init__.py
@@ -16,7 +16,7 @@ from pkg_resources import DistributionNotFound, get_distribution
 
 
 try:
-    __version__ = get_distribution('TemplateDemo').version
+    __version__ = get_distribution('pytest-expecter').version
 except DistributionNotFound:
     __version__ = '(local)'
 
@@ -111,9 +111,21 @@ class expect:
     def __repr__(self):
         return 'expect(%s)' % repr(self._actual)
 
+    def is_(self, other):
+        """Ensure that ``other`` is identical to the actual value."""
+        __tracebackhide__ = _hidetraceback()  # pylint: disable=unused-variable
+        label = 'condition' if isinstance(self._actual, bool) else 'value'
+        assert self._actual is other, "Expected %s to be %s, but it was %s" % (
+            label,
+            repr(other),
+            repr(self._actual),
+        )
+        return self
+
     def isinstance(self, expected_cls):
-        """
-        Ensures that the actual value is of type ``expected_cls`` (like ``assert isinstance(actual, MyClass)``).
+        """Ensure the actual value is of type ``expected_cls``.
+
+        This is similar to ``assert isinstance(actual, MyClass)``.
         """
         __tracebackhide__ = _hidetraceback()  # pylint: disable=unused-variable
         if isinstance(expected_cls, tuple):
@@ -127,8 +139,9 @@ class expect:
         )
 
     def contains(self, other):
-        """
-        Ensure that ``other`` is in the actual value (like ``assert other in actual``).
+        """Ensure that ``other`` is in the actual value.
+
+        This is similar to ``assert other in actual``.
         """
         __tracebackhide__ = _hidetraceback()  # pylint: disable=unused-variable
 
@@ -146,9 +159,7 @@ class expect:
         assert other in self._actual, msg
 
     def icontains(self, other):
-        """
-        Same as ``contains` but ignoring case.
-        """
+        """Same as ``contains` but ignoring case."""
         __tracebackhide__ = _hidetraceback()  # pylint: disable=unused-variable
 
         if isinstance(self._actual, str) and '\n' in self._actual:
@@ -167,10 +178,25 @@ class expect:
 
         assert expected in actual, msg
 
+    def includes(self, other):
+        """Same as ``contains`` but with alternate phrasing."""
+        __tracebackhide__ = _hidetraceback()  # pylint: disable=unused-variable
+
+        if isinstance(self._actual, str) and '\n' in self._actual:
+            msg = "Given text:\n\n%s\n\nExpected to include %s but didn't" % (
+                self._actual.strip(),
+                repr(other),
+            )
+        else:
+            msg = "Expected %s to include %s but it didn't" % (
+                repr(self._actual),
+                repr(other),
+            )
+
+        assert other in self._actual, msg
+
     def does_not_contain(self, other):
-        """
-        Opposite of ``contains``
-        """
+        """Opposite of ``contains``."""
         __tracebackhide__ = _hidetraceback()  # pylint: disable=unused-variable
 
         if isinstance(self._actual, str) and '\n' in self._actual:
@@ -187,9 +213,7 @@ class expect:
         assert other not in self._actual, msg
 
     def excludes(self, other):
-        """
-        Opposite of ``contains`` with alternate phrasing.
-        """
+        """Opposite of ``contains`` with alternate phrasing."""
         __tracebackhide__ = _hidetraceback()  # pylint: disable=unused-variable
 
         if isinstance(self._actual, str) and '\n' in self._actual:
@@ -206,9 +230,7 @@ class expect:
         assert other not in self._actual, msg
 
     def iexcludes(self, other):
-        """
-        Same as ``excludes`` but ignoring case.
-        """
+        """Same as ``excludes`` but ignoring case."""
         __tracebackhide__ = _hidetraceback()  # pylint: disable=unused-variable
 
         if isinstance(self._actual, str) and '\n' in self._actual:
@@ -228,9 +250,7 @@ class expect:
         assert expected not in actual, msg
 
     def startswith(self, other):
-        """
-        Ensure that ``other`` starts the actual value.
-        """
+        """Ensure that ``other`` starts the actual value."""
         __tracebackhide__ = _hidetraceback()  # pylint: disable=unused-variable
 
         if isinstance(self._actual, str) and '\n' in self._actual:
@@ -247,9 +267,7 @@ class expect:
         assert self._actual.startswith(other), msg
 
     def istartswith(self, other):
-        """
-        Same as ``startswith`` but ignoring case.
-        """
+        """Same as ``startswith`` but ignoring case."""
         __tracebackhide__ = _hidetraceback()  # pylint: disable=unused-variable
 
         if isinstance(self._actual, str) and '\n' in self._actual:
@@ -266,9 +284,7 @@ class expect:
         assert self._actual.lower().startswith(other.lower()), msg
 
     def endswith(self, other):
-        """
-        Ensure that ``other`` ends the actual value.
-        """
+        """Ensure that ``other`` ends the actual value."""
         __tracebackhide__ = _hidetraceback()  # pylint: disable=unused-variable
 
         if isinstance(self._actual, str) and '\n' in self._actual:
@@ -285,9 +301,7 @@ class expect:
         assert self._actual.endswith(other), msg
 
     def iendswith(self, other):
-        """
-        Same as ``endswith`` but ignoring case.
-        """
+        """Same as ``endswith`` but ignoring case."""
         __tracebackhide__ = _hidetraceback()  # pylint: disable=unused-variable
 
         if isinstance(self._actual, str) and '\n' in self._actual:
@@ -305,7 +319,7 @@ class expect:
 
     @staticmethod
     def raises(expected_cls=Exception, message=None):
-        """Ensure that an exception is raised. E.g.,
+        """Ensure that an exception is raised.
 
         ::
 
@@ -391,7 +405,7 @@ class _CustomExpectation:
     def enforce(self, *args, **kwargs):
         __tracebackhide__ = _hidetraceback()  # pylint: disable=unused-variable
         if not self._predicate(self._actual, *args, **kwargs):
-            predicate_name = self._predicate.__name__
+            predicate_name = self._predicate.__name__.replace('_', ' ')
             raise AssertionError(
                 'Expected that %s %s, but %s'
                 % (repr(self._actual), predicate_name, self._negative_verb())